adding menu repository

+++ src/main/java/menu/model/beer/Beer.groovy
@@ -14,7 +14,7 @@ public class Beer {
 
     @Id
     @GeneratedValue
-    Integer id
+    long id
     String name
